+++ apps/www/src/app/(main)/(docs)/components/circular-progress/examples/colors.tsx
@@ -2,6 +2,7 @@ import { CircularProgress, type CircularProgressProps } from "@jamsr-ui/react";
 
 export const CircularProgressColors = () => {
   const colors: CircularProgressProps["color"][] = [
+    "default",
     "primary",
     "secondary",
     "success",