+++ apps/www/src/app/(main)/(docs)/components/circular-progress/examples/with-label.tsx
@@ -1,0 +1,5 @@
+import { CircularProgress } from "@jamsr-ui/react";
+
+export const CircularProgressLabel = () => {
+  return <CircularProgress value={50} showLabel isIntermediate={false} />;
+};