+++ apps/www/src/app/(main)/(docs)/components/circular-progress/examples/with-value.tsx
@@ -1,5 +1,0 @@
-import { CircularProgress } from "@jamsr-ui/react";
-
-export const CircularProgressValue = () => {
-  return <CircularProgress value={50} />;
-};