+++ packages/components/charts/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/chart
 
+## 0.0.2
+
+### Patch Changes
+
+- feat: circular-progress
+
 ## 0.0.1
 
 ### Patch Changes