+++ packages/components/rhf/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @jamsr-ui/rhf
 
+## 0.0.61
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/file-upload-multi@0.0.103
+  - @jamsr-ui/file-upload-single@0.0.106
+  - @jamsr-ui/editor@0.0.117
+  - @jamsr-ui/input@0.0.107
+  - @jamsr-ui/select@0.0.103
+  - @jamsr-ui/autocomplete@0.0.116
+  - @jamsr-ui/textarea@0.0.107
+  - @jamsr-ui/tags-input@0.0.45
+
 ## 0.0.60
 
 ### Patch Changes