+++ packages/components/select/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @jamsr-ui/select
 
+## 0.0.103
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/popover@0.0.80
+  - @jamsr-ui/button@0.0.84
+  - @jamsr-ui/chip@0.0.92
+
 ## 0.0.102
 
 ### Patch Changes